+++ tabletest-reporter-junit/src/main/java/io/github/nchaugen/tabletest/reporter/junit/CellRole.java
@@ -17,5 +17,7 @@ package io.github.nchaugen.tabletest.reporter.junit;
 
 public enum CellRole {
     EXPECTATION,
-    SCENARIO
+    SCENARIO,
+    PASSED,
+    FAILED
 }